Make probing a public package.
