+++ go.mod
@@ -2,7 +2,10 @@ module package-operator.run
 
 go 1.20
 
-replace package-operator.run/apis => ./apis
+replace (
+	package-operator.run/apis => ./apis
+	package-operator.run/pkg => ./pkg
+)
 
 require (
 	github.com/Masterminds/sprig/v3 v3.2.3
@@ -10,7 +13,6 @@ require (
 	github.com/disiqueira/gotree v1.0.0
 	github.com/go-logr/logr v1.2.4
 	github.com/gobwas/glob v0.2.3
-	github.com/google/cel-go v0.16.1
 	github.com/google/go-containerregistry v0.16.1
 	github.com/magefile/mage v1.15.0
 	github.com/mt-sre/devkube v0.7.1
@@ -35,6 +37,7 @@ require (
 	k8s.io/kube-openapi v0.0.0-20231010175941-2dd684a91f00
 	k8s.io/utils v0.0.0-20230726121419-3b25d923346b
 	package-operator.run/apis v1.8.0
+	package-operator.run/pkg v0.0.0-00010101000000-000000000000
 	sigs.k8s.io/controller-runtime v0.16.3
 	sigs.k8s.io/kind v0.20.0
 	sigs.k8s.io/yaml v1.3.0
@@ -75,6 +78,7 @@ require (
 	github.com/gogo/protobuf v1.3.2 // indirect
 	github.com/golang/groupcache v0.0.0-20210331224755-41bb18bfe9da // indirect
 	github.com/golang/protobuf v1.5.3 // indirect
+	github.com/google/cel-go v0.16.1 // indirect
 	github.com/google/gnostic-models v0.6.8 // indirect
 	github.com/google/go-cmp v0.6.0 // indirect
 	github.com/google/gofuzz v1.2.0 // indirect
