+++ internal/controllers/phase_reconciler.go
@@ -26,7 +26,7 @@ import (
 	corev1alpha1 "package-operator.run/apis/core/v1alpha1"
 	manifestsv1alpha1 "package-operator.run/apis/manifests/v1alpha1"
 	"package-operator.run/internal/preflight"
-	"package-operator.run/internal/probing"
+	"package-operator.run/pkg/probing"
 )
 
 // PhaseReconciler reconciles objects within a ObjectSet phase.
