+++ pkg/probing/doc.go
@@ -1,0 +1,3 @@
+// The probing package contains declarative status probing implementations
+// for various different Kubernetes objects.
+package probing
